Look-up token types by Cosmos min-denom
